% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_sat_data.R
\name{collect_ancillary_data}
\alias{collect_ancillary_data}
\title{Collect point based satellite}
\usage{
collect_ancillary_data(df, source = tempdir(), dest)
}
\arguments{
\item{df}{a data frame with coordinates}

\item{source}{a source where to download and or source data from}

\item{dest}{where to store the final results (by default results) are
also returned to the workspace}
}
\value{
a
}
\description{
Give a data frame with coordinates
this routine collects all data ancillary
satellite data for ML model development.
This includes various vegetation indices,
meteo data, etc.
}
