% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_gee_data.R
\name{collect_gee_data}
\alias{collect_gee_data}
\title{Collect point based satellite}
\usage{
collect_gee_data(
  df,
  source = "/scratch/LACUNA/remote_sensing/",
  dest = "/scratch/LACUNA/data_product/remote_sensing_data/"
)
}
\arguments{
\item{df}{a data frame with coordinates}

\item{source}{a source where to download and or source data from}

\item{dest}{where to store the final results (by default results) are
also returned to the workspace}
}
\value{
a
}
\description{
Give a data frame with coordinates
this routine collects all data ancillary
satellite data for ML model development.
This includes various vegetation indices,
meteo data, etc. Only repos on GEE are evaluated
}
