% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_image_list.R
\name{generate_image_list}
\alias{generate_image_list}
\title{List all valid images}
\usage{
generate_image_list(
  path = "/backup/see_it_grow",
  season_filter = "LR2021",
  ml_labels = TRUE
)
}
\arguments{
\item{path}{where to store the full list as csv file}

\item{season_filter}{what season to filter out (LR2021 by default)}
}
\value{

}
\description{
All valid images for processing, as
manually screened by surveyors
}
