% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_labels.R
\name{ml_labels}
\alias{ml_labels}
\title{Return ML labels}
\usage{
ml_labels(path = "data/ml_labels/ml/")
}
\arguments{
\item{path}{path to the various ML prediction files}
}
\value{

}
\description{
Return precomputed ML labels
for all images (all seasons)
}
