% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_spatial_info.R
\name{pbi_map_gadm}
\alias{pbi_map_gadm}
\title{Maps locations to town centroids}
\usage{
pbi_map_gadm(df, country)
}
\arguments{
\item{df}{data frame containing lat lon locations}

\item{country}{country to use in mapping locations to town centroids}
}
\value{
dataframe with lat lon values mapped to centroids
}
\description{
For a particular country maps locations of fields to the
centroid of towns. This anonymizes the data by obfuscating
the original location.
}
