% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_spatial_info.R
\name{pbi_map_grid}
\alias{pbi_map_grid}
\title{Maps coordinates to grid centroids}
\usage{
pbi_map_grid(df, resolution = 2.5)
}
\arguments{
\item{df}{data frame containing lat lon locations}

\item{resolution}{resolution of the map in minutes (2.5, 5, 10)}
}
\value{
dataframe with lat lon values mapped to centroids
}
\description{
Anonymizes lat/lon coordinates by remapping them
to low resolution grid cell centroids
}
