% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_crop_cuts.R
\name{process_crop_cuts}
\alias{process_crop_cuts}
\title{Format crop cut XLSX data}
\usage{
process_crop_cuts(
  xls = "/backup/see_it_grow/LR2021/Reports/LR2021 CCE data sheet - PBI & IUK.xlsx"
)
}
\arguments{
\item{xls}{xlsx file with the crop cut data}
}
\value{

}
\description{
Custom function to transform the XLS data
into a tidy data frame for further processing
}
