% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_details.R
\name{site_details}
\alias{site_details}
\title{Collect site details}
\usage{
site_details(site_list = FALSE, path)
}
\arguments{
\item{site_list}{}

\item{path}{}
}
\value{

}
\description{
Function to process ancillary data
into a tidy data frame, this includes
soil types, yields etc.
}
